+++ CHANGELOG.md
@@ -42,6 +42,9 @@ and this project adheres to
 - Fixed bug where the variable in an `"ARGS"` binding couldn't be referenced
   later in the binding list.
 
+- Fixed incorrect error message when calling a varargs function with too few
+  arguments.
+
 ## [1.2] (December 24, 2025)
 
 ### Added